requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
transformers==4.45.0
torch>=2.1.0
sentence-transformers==3.0.0
faiss-cpu==1.13.0
requests==2.32.3
pydantic==2.8.0
